continuation squellette

+++ Game.java
@@ -2,7 +2,7 @@ import Entity.Player.Octaling;
 
 public class Game {
     public static void main(String[] args){
-        Octaling octo = new Octaling(0, 0, 0, null, 0, 0, 0);
+        Octaling octo = new Octaling((short)0, (short)0, (short)0, null, (short)0, (short)0, 0);
         octo.print_info();
     }
 }
